moves Dropdown to Controls folder
